+++ api/.env
@@ -1,0 +1,7 @@
+NODE_ENV=development
+
+DB_HOST=localhost
+DB_USER=root
+DB_PASSWORD=
+DB_PORT=3306
+DB_NAME=xendit-test
