+++ api/src/db/models/customer.js
@@ -3,7 +3,7 @@ const {
   Model
 } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-  class ProductType extends Model {
+  class Customer extends Model {
     /**
      * Helper method for defining associations.
      * This method is not a part of Sequelize lifecycle.
@@ -13,16 +13,21 @@ module.exports = (sequelize, DataTypes) => {
       // define association here
     }
   }
-  ProductType.init({
-    product_type_id: {
+  Customer.init({
+    customer_id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
       primaryKey: true
     },
-    name: DataTypes.STRING
+    name: DataTypes.STRING,
+    contact_number: DataTypes.STRING
   }, {
     sequelize,
-    modelName: 'ProductType',
+    modelName: 'Customer',
+    tableName: 'customer',
+    timestamps: false,
+    createdAt: false,
+    updatedAt: false,
   });
-  return ProductType;
+  return Customer;
 };
