--- backend/config/config.json
@@ -1,23 +1,0 @@
-{
-  "development": {
-    "username": "root",
-    "password": "",
-    "database": "xendit-test",
-    "host": "localhost",
-    "dialect": "mysql"
-  },
-  "test": {
-    "username": "root",
-    "password": "",
-    "database": "xendit-test",
-    "host": "localhost",
-    "dialect": "mysql"
-  },
-  "production": {
-    "username": "root",
-    "password": "",
-    "database": "xendit-test",
-    "host": "localhost",
-    "dialect": "mysql"
-  }
-}
