--- backend/index.ts
@@ -1,8 +1,0 @@
-import express from "express";
-
-const app = express();
-const port = 8080;
-
-app.listen(port, () => {
-  console.log(`[server] server started at http://localhost:${port}`);
-});
