--- backend/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "fullstack-coding-test-backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "tsc && node compiled/index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@sequelize/core": "^7.0.0-alpha.10",
-    "@types/express": "^4.17.13",
-    "@types/node": "^17.0.21",
-    "express": "^4.17.3",
-    "mysql2": "^2.3.3",
-    "pg": "^8.0.0",
-    "pg-hstore": "^2.3.4",
-    "sequelize": "^6.17.0"
-  },
-  "devDependencies": {
-    "sequelize-cli": "^6.4.1",
-    "typescript": "^4.6.2"
-  }
-}
